#205
- media-gallery: вынесен в отдельный компонент
- hooks: добавлен useMounted
