+++ src/common/components/media-gallery/index.ts
@@ -1,0 +1,2 @@
+export type { MediaGalleryProps, MediaGalleryStyle } from './types';
+export { MediaGallery, MediaSlide } from './media-gallery';
