+++ src/common/components/media-gallery/media-gallery.module.scss
