+++ src/common/components/media-gallery/types.ts
@@ -1,0 +1,15 @@
+import type { CSSProperties, ReactNode } from 'react';
+
+export interface MediaGalleryStyle extends CSSProperties {
+  '--media-gallery-width'?: string;
+  '--media-gallery-height'?: string;
+}
+
+export interface MediaGalleryProps {
+  style?: MediaGalleryStyle;
+  className?: string;
+  children: ReactNode;
+  targetIndex?: number;
+  onChangeTargetIndex?: (newIndex: number) => void;
+  loading?: boolean;
+}
