+++ src/common/components/media-modal/__stories__/08-gallery.stories.tsx
@@ -6,15 +6,13 @@ import {
   MediaFooter,
   Thumbnails,
   Thumbnail,
-  MediaGallery,
   MediaView,
-  MediaSlide,
   MediaArea,
 } from '@sima-land/moleculas/common/components/media-modal';
-import { CSSProperties, Fragment, useRef, useState } from 'react';
+import { MediaGallery, MediaSlide } from '@sima-land/moleculas/common/components/media-gallery';
+import { Fragment, useState } from 'react';
 import { Modal } from '@sima-land/ui-nucleons/modal';
 import { mixed } from '../__mocks__';
-import { usePageScrollLock } from '@sima-land/ui-nucleons/_internal/page-scroll-lock';
 
 export default {
   title: 'common/MediaLayout',
@@ -23,47 +21,6 @@ export default {
     layout: 'fullscreen',
   },
 };
-
-export function ExampleCleanGallery() {
-  const [targetIndex, setTargetIndex] = useState(0);
-
-  const ref = useRef<HTMLDivElement>(null);
-  usePageScrollLock(ref);
-
-  const styles = {
-    root: {
-      position: 'fixed',
-      display: 'flex',
-      justifyContent: 'center',
-      '--media-gallery-width': '100vw',
-      '--media-gallery-height': '100vh',
-      '--media-view-width': 'var(--media-gallery-width)',
-      '--media-view-height': 'var(--media-gallery-height)',
-    } as any,
-    item: {
-      width: 'var(--media-gallery-width)',
-      height: 'var(--media-gallery-height)',
-      background: '#ddd',
-      display: 'flex',
-      alignItems: 'center',
-      justifyContent: 'center',
-    } satisfies CSSProperties,
-  };
-
-  return (
-    <div ref={ref} style={styles.root}>
-      <div style={{ background: '#000' }}>
-        <MediaGallery targetIndex={targetIndex} onChangeTargetIndex={setTargetIndex}>
-          {mixed.map((item, index) => (
-            <MediaSlide key={index}>
-              <MediaView media={item} />
-            </MediaSlide>
-          ))}
-        </MediaGallery>
-      </div>
-    </div>
-  );
-}
 
 export function ExampleGallery() {
   const [targetIndex, setTargetIndex] = useState(0);
@@ -121,7 +78,7 @@ export function ExampleGallery() {
                   {item.type === 'image' && (
                     <Thumbnail
                       type='preview-image'
-                      src={item.data.src}
+                      src={item.data.thumbnail}
                       checked={targetIndex === index}
                       onClick={() => setTargetIndex(index)}
                     />
