+++ src/common/components/media-modal/__stories__/09-gallery-video.stories.tsx
@@ -6,11 +6,10 @@ import {
   MediaFooter,
   Thumbnails,
   Thumbnail,
-  MediaGallery,
   MediaView,
   MediaArea,
-  MediaSlide,
 } from '@sima-land/moleculas/common/components/media-modal';
+import { MediaGallery, MediaSlide } from '../../media-gallery';
 import { useState } from 'react';
 import { Modal } from '@sima-land/ui-nucleons/modal';
 import { videos } from '../__mocks__';
