+++ src/common/components/media-modal/__stories__/loading-mixed.stories.tsx
@@ -8,7 +8,6 @@ import {
   HeaderLayout,
   Thumbnails,
   Thumbnail,
-  MediaGallery,
 } from '@sima-land/moleculas/common/components/media-modal';
 import { Modal } from '@sima-land/ui-nucleons/modal';
 import { Tabs } from '@sima-land/ui-nucleons/tabs';
@@ -45,9 +44,7 @@ export function LoadingMixed() {
             </HeaderLayout>
           </MediaHeader>
 
-          <MediaMain>
-            <MediaGallery loading />
-          </MediaMain>
+          <MediaMain></MediaMain>
 
           <MediaAside>
             <Thumbnails>
