+++ src/common/components/media-modal/__stories__/loading-video.stories.tsx
@@ -8,7 +8,6 @@ import {
   HeaderLayout,
   Thumbnails,
   Thumbnail,
-  MediaView,
   ProductBrief,
 } from '@sima-land/moleculas/common/components/media-modal';
 import { Modal } from '@sima-land/ui-nucleons/modal';
@@ -44,9 +43,7 @@ export function LoadingVideo() {
             </HeaderLayout>
           </MediaHeader>
 
-          <MediaMain>
-            <MediaView loading />
-          </MediaMain>
+          <MediaMain></MediaMain>
 
           <MediaAside>
             <Thumbnails>
