+++ src/common/components/media-modal/index.ts
@@ -3,16 +3,12 @@ import { Preset, HeaderLayout } from './parts';
 import { ScrollSection } from './parts/scroll-section';
 import { Thumbnail, ThumbnailProps } from './parts/thumbnail';
 import { Thumbnails, ThumbnailsProps } from './parts/thumbnails';
-import { MediaGalleryProps, MediaGallery, MediaSlide } from './parts/media-gallery';
 import { MediaViewProps, MediaView } from './parts/media-view';
 import { ReviewProps, Review } from './parts/review';
 import { ProductBriefProps, ProductBrief } from './parts/product-brief';
 import { MediaArea } from './parts/media-area';
 
 export type {
-  // main content - multiple
-  MediaGalleryProps,
-
   // main content - single
   MediaViewProps,
 
@@ -33,10 +29,6 @@ export {
   MediaAside,
   MediaFooter,
   MediaArea,
-
-  // main content - multiple
-  MediaGallery,
-  MediaSlide,
 
   // main content - single
   MediaView,
