+++ src/common/components/media-modal/parts/all-round-view.tsx
@@ -5,7 +5,7 @@ import {
   AllRoundViewProps,
   ViewState,
 } from '../../../../desktop/components/gallery-modal/components/all-round-view';
-import { useMounted } from '../utils';
+import { useMounted } from '../../../../common/hooks';
 import TouchSVG from '@sima-land/ui-quarks/icons/64x64/Colorful/TouchScrollX';
 import styles from './all-round-view.module.scss';
