+++ src/common/components/media-modal/utils.ts
@@ -96,21 +96,6 @@ export function useClientRect(ref: RefObject<HTMLElement>) {
 }
 
 /**
- * Возвращает true если компонент смонтирован.
- * @return True если компонент смонтирован.
- */
-export function useMounted() {
-  const [mounted, setMounted] = useState(false);
-
-  useIsomorphicLayoutEffect(() => {
-    setMounted(true);
-    return () => setMounted(false);
-  }, []);
-
-  return mounted;
-}
-
-/**
  * Хук получения доступа к visualViewport.
  * @param handle Получит visualViewport.
  */
