+++ src/common/hooks/index.ts
@@ -26,3 +26,18 @@ export function useImageStub(src?: string, onError?: ReactEventHandler<HTMLImage
 
   return { failed, setFailed, handleError };
 }
+
+/**
+ * Возвращает true если компонент смонтирован.
+ * @return True если компонент смонтирован.
+ */
+export function useMounted() {
+  const [mounted, setMounted] = useState(false);
+
+  useIsomorphicLayoutEffect(() => {
+    setMounted(true);
+    return () => setMounted(false);
+  }, []);
+
+  return mounted;
+}
